frontend-tfserving/requirements.txt
pillow
requests
numpy
pandas
scikit-learn==1.0.2
streamlit
typing_extensions
tensorflow-cpu



